Utilities

Test Videos

A small number of test videos are available through VideoIO.TestVideos. These are short videos in a variety of formats with non-restrictive (public domain or Creative Commons) licenses.

VideoIO.TestVideos.available — Function

available()

Print a list of all available test videos.

VideoIO.TestVideos.testvideo — Function

testvideo(name, ops...)

Returns an AVInput object for the given video name. The video will be downloaded if it isn't available.

VideoIO.TestVideos.download_all — Function

download_all()

Downloads all test videos.

VideoIO.TestVideos.remove_all — Function

remove_all()

Remove all test videos.